Business people management bug fixes

See merge request urs/urs-frontend!26

+++ src/containers/DeleteBusinessAttachment.tsx
@@ -1,17 +1,17 @@
-import Loader from '@/components/Loader';
-import Modal from '@/components/Modal';
-import Button from '@/components/inputs/Button';
-import { useDeleteBusinessAttachmentMutation } from '@/states/api/coreApiSlice';
+import Loader from "@/components/Loader";
+import Modal from "@/components/Modal";
+import Button from "@/components/inputs/Button";
+import { useDeleteBusinessAttachmentMutation } from "@/states/api/coreApiSlice";
 import {
   removeBusinessAttachment,
   setDeleteBusinessAttachmentModal,
-} from '@/states/features/businessSlice';
-import { AppDispatch, RootState } from '@/states/store';
-import { useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { useDispatch } from 'react-redux';
-import { ErrorResponse } from 'react-router-dom';
-import { toast } from 'react-toastify';
+} from "@/states/features/businessSlice";
+import { AppDispatch, RootState } from "@/states/store";
+import { useEffect } from "react";
+import { useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
+import { ErrorResponse } from "react-router-dom";
+import { toast } from "react-toastify";
 
 const DeleteBusinessAttachment = () => {
   // STATE VARIABLES
@@ -27,18 +27,20 @@ const DeleteBusinessAttachment = () => {
       isSuccess: deleteBusinessAttachmentIsSuccess,
       isError: deleteBusinessAttachmentIsError,
       error: deleteBusinessAttachmentError,
+      reset: resetDeleteBusinessAttachment,
     },
   ] = useDeleteBusinessAttachmentMutation();
 
   // HANDLE DELETE BUSINESS ATTACHMENT RESPONSE
   useEffect(() => {
-    if (deleteBusinessAttachmentIsSuccess) {
+    if (deleteBusinessAttachmentIsSuccess && deleteBusiessAttachmentModal) {
       dispatch(removeBusinessAttachment(selectedBusinessAttachment));
       dispatch(setDeleteBusinessAttachmentModal(false));
+      resetDeleteBusinessAttachment();
     } else if (deleteBusinessAttachmentIsError) {
       const errorResponse =
         (deleteBusinessAttachmentError as ErrorResponse)?.data?.message ||
-        'An error occurred while deleting business attachment. Refresh the page and try again.';
+        "An error occurred while deleting business attachment. Refresh the page and try again.";
       toast.error(errorResponse);
     }
   }, [
@@ -47,6 +49,8 @@ const DeleteBusinessAttachment = () => {
     deleteBusinessAttachmentError,
     dispatch,
     selectedBusinessAttachment,
+    deleteBusiessAttachmentModal,
+    resetDeleteBusinessAttachment,
   ]);
 
   return (
@@ -54,35 +58,37 @@ const DeleteBusinessAttachment = () => {
       isOpen={deleteBusiessAttachmentModal}
       onClose={() => {
         dispatch(setDeleteBusinessAttachmentModal(false));
+        resetDeleteBusinessAttachment();
       }}
     >
-      <section className="w-full flex flex-col gap-3">
-        <h1 className="text-red-600 uppercase font-medium">
+      <section className="flex flex-col w-full gap-3">
+        <h1 className="font-medium text-red-600 uppercase">
           Delete "{selectedBusinessAttachment?.fileName}"
         </h1>
         <p>
-          Are you sure you want to delete{' '}
+          Are you sure you want to delete{" "}
           <span className="font-medium">
             "{selectedBusinessAttachment?.fileName}"
-          </span>{' '}
-          of{' '}
+          </span>{" "}
+          of{" "}
           <span className="font-medium">
             {selectedBusinessAttachment?.attachmentType}
           </span>
           ?
         </p>
         <p>This action cannot be undone!</p>
-        <menu className="w-full flex items-center gap-3 justify-between mt-4">
+        <menu className="flex items-center justify-between w-full gap-3 mt-4">
           <Button
-            value={'Cancel'}
+            value={"Cancel"}
             onClick={(e) => {
               e.preventDefault();
               dispatch(setDeleteBusinessAttachmentModal(false));
+              resetDeleteBusinessAttachment();
             }}
           />
           <Button
             danger
-            value={deleteBusinessAttachmentIsLoading ? <Loader /> : 'Delete'}
+            value={deleteBusinessAttachmentIsLoading ? <Loader /> : "Delete"}
             onClick={(e) => {
               e.preventDefault();
               deleteBusinessAttachment({ id: selectedBusinessAttachment?.id });
